+++ datamodel/app/view/vw_tww_measurement_series.py
